+++ src/Componets/reactHooks/UseContex.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useContext } from 'react';
-import { Context } from '../App.jsx';
+import { Context } from '../../Data/context.jsx';
 import UseContextChild from './UseContextChild';
 function UseContex() {
   const userInfo = useContext(Context);
