+++ src/Data/context.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+export const Context = createContext('Default Value');
+export const workContext = createContext('Default Value');
